logic: splitting the validate funcs to match the convention
