+++ azurerm/internal/services/logic/validate/integration_account_id.go
@@ -1,0 +1,21 @@
+package validate
+
+import (
+	"fmt"
+
+	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/logic/parse"
+)
+
+func IntegrationAccountID(i interface{}, k string) (warnings []string, errors []error) {
+	v, ok := i.(string)
+	if !ok {
+		errors = append(errors, fmt.Errorf("expected type of %q to be string", k))
+		return warnings, errors
+	}
+
+	if _, err := parse.IntegrationAccountID(v); err != nil {
+		errors = append(errors, fmt.Errorf("Can not parse %q as a Integration Account resource id: %v", k, err))
+	}
+
+	return warnings, errors
+}
